delete shadowcopy signature and ie martian signature
To the best of my knowledge the term "martian" was originally coined by
Will Metcalf in his signatures. Although just at the start of its
possibilities, this Signature provides a real quick assessment of the
websites loaded - does it drop any malicious payload or not? (Note that
not all payloads require a new process to be started).