Signature to detect document files creating network traffic
This is my first conversion sig to 2.0 so please check although it appears to work fine and is the start of me trying to convert some of the signatures from cuckoo-modified over into cuckoo-2.0 so at least a start has been made. The signature is to convert the ideas from this signature:
https://raw.githubusercontent.com/spender-sandbox/community-modified/master/modules/signatures/network_docfile_http.py

If the above link can be used to verifiy the logic of this signature before commit then that would be best given this is my first cuckoo 2.0 signature and also my pythons skills are fine but needs an expert eye :-) I intend to start the conversion process as I can because as previously discussed in general there is a lot of excellent signatures and while I may not manage everything at least it helps out with what I can.