Create another ransomware file modification detection
This is based on large numbers of unknown files being dropped to disk. It is based on this https://raw.githubusercontent.com/spender-sandbox/community-modified/master/modules/signatures/ransomware_filemodifications.py; specifically this bit:

        if "dropped" in self.results:
            droppedunknowncount = 0
            for dropped in self.results["dropped"]:
                mimetype = dropped["type"]
                filename = dropped["name"]
                if mimetype == "data" and ".tmp" not in filename:
                    droppedunknowncount += 1            
            if droppedunknowncount > 50:
                self.data.append({"drops_unknown_mimetypes" : "Drops %s unknown file mime types which may be indicative of encrypted files being written back to disk" % (droppedunknowncount)})
                ret = True 

I wanted it to take the value in droppedcount and update it so it would be a message like "Drops 3243 unknown file mime types indicative of ransomware writing encrypted files back to disk" with the 3243 but for some reason this wouldn't take as I expected so I have left it as this for now. With this and the filemove signatures however it seems to cover most ransomware file modification activity.