Move over office EXE dropper sig
This is a conversion of the sig https://raw.githubusercontent.com/spender-sandbox/community-modified/master/modules/signatures/office_write_exe.py from Will Metcalf. Tested and works against a recent Dridex email from late last week; this sig also detects cases where the EXE is embedded in the office file.

The only bit that seemed to have issue taking in the conversion was buff[0:1] == "MZ". I just made it "MZ" in buff as I felt for now there is enough conditions and it works but if it can be determined why that would't take it would make the sig more accurate and true to Will Metcalf's original.