Create new sig for using computer details in c2
This sig simply started when looking at Dridex. The intention is that various bits of information used in C2 construction for identification of the computer when they are queried can then be looked for in Crypto calls (for hashing the data before sending it as malware does) or directly in network traffic.

Right now it is only looking for username & computer name but this could be expanded to include checks for:
- IP Address
- MAC Address
- AV installed?
- Product ID
And so on.

Still it could be grown into a useful sig for identifying how bits of information from the computer are then used in C2 although it will take some research & testing to implement further checks.