Update persistance locations and methods
I originally was not going to sort out the signatures that exist between cuckoo-modified & cuckoo 2.0 until the based signatures were the same. However as I was having problems with this not firing I have updated it from this:

https://raw.githubusercontent.com/spender-sandbox/community-modified/master/modules/signatures/persistence_autorun.py

I have not included the whitelist that exists on the cuckoo-modified signature and I am aware some of the API stuff overlaps the service creation signature but the main check is the start type. I was going to just update the service creation signature but I decided against it as I would rather present service creation for autoruns showing with the rest of them. However it does result in a weird combined registry/API call results when the self.marks are reported. Still it looks ok and communicates this is persitance fine & these updates also fix the issues with the Wow6432Node keys not being fired on.