Add in signature recon_beacon
This is a converted version of this signature: https://github.com/spender-sandbox/community-modified/pull/140. This was actually pointed out to me as being present and so an update to the original signature on cuckoo-modified has been submitted. 

This specific signature is closer to my original signature though which I submitted and closed here simply because this is not possible and so globally just getting the username and computername rather than relying on checking APIs is not possible

        initproc = self.get_initial_process()
        uname = self.get_environ_entry(initproc, "UserName")
        cname = self.get_environ_entry(initproc, "ComputerName")