Add in bootkit signature
Converted cuckoomodified signature & I have tested this signature using Petya ransomware.https://blog.malwarebytes.com/threat-analysis/2016/04/petya-ransomware/

ORIGINAL CUCKOO MODIFIED SIGNATURE: https://github.com/spender-sandbox/community-modified/blob/master/modules/signatures/bootkit.py

Also for the API call NtDuplicateObject I did not have an example call to do the conversion with so I guessed TargetHandle & SourceHandle on cuckoo-modified/1.3 becomes target_handle and source_handle as this is the typical conversion but if this is incorrect could it please be set as the correct argument name?

Also another thing I noted during the analysis was the API NtRaiseHardError which is used by Petya to reboot the system is not shown in the behaviour logs and I am wondering if this is hooked as it is used in the signature (revised one I submitted) for detecting malware restarting sandbox. I can see the call just before it LdrGetProcedureAddress for NtRaiseHardError but that is the last API I have in cuckoo 2.0 suggestng NtRaiseHardError is not hooked.