Create signature for self deletion
The concept of this signature is from the cuckoo-modified signature https://github.com/spender-sandbox/community-modified/blob/master/modules/signatures/deletes_self.py.

However I have rewritten its operation due to some differences so the way this is detected is not achieved the same way as the original signature so I have also put my name on this due to it being more than a simple signature upgrade. The logic is basically put all processes into a list and then check if a deleted file exists in position 0 (first process) for a match.

It has been tested on several malware samples which deletes their original binary such as TeslaCrypt and it works fine.