Add in signature for creating ransomware message
This is a conversion of this signature I put together for cuckoo-modified https://github.com/spender-sandbox/community-modified/pull/157. The concerns that were raised during my initial submission of that signature around performance for looking at every NTWriteFile call stands for this signature. On the other one I got round the performance issues by:

- For NTWriteFile only looking for bootkit modifications
- For files actually written to disk looking through dropped files as cuckoo-modified shows the first 4KB.
- NtWriteFile doesn't actually show the written file name it seems in the cuckoo 2.0 reports, just buffer and filehandle.

With this one however I am closer to my original sig but with some performance and accuracy improvements but I have tested with ransomware & it works fine.