Extract process name
Hey, 

Thanks for this. I have been looking for a way to do a sig to detect heap sprays reliably :)

I have added in extracton of the process name to make it easier ot identify what is being exploited. It may also be useful to whitelist executable content so only things like browsers, PDF readers, office etc is covered which are more likely to actually be subject to heap sprays and also reduce FP chances. If it can be made that the package used it extracted on automatic submission too then something like it not being: self.results["info"]["package"] not in ["exe", "rar", "zip", "dll", "regsvr"]: could be used to do this.

On the exploit front I would like to convert the following but require changes if you wanted to further increase exploit detection :-D

- Stack Pivot (requires stuff from cuckoosploit. I see pull request submitted https://github.com/cuckoosandbox/monitor/pull/17): https://github.com/spender-sandbox/community-modified/blob/master/modules/signatures/stack_pivot.py
- DEP_Bypass (APIs in cuckoo 2.0 don't seem to show old protection): https://github.com/spender-sandbox/community-modified/blob/master/modules/signatures/dep_bypass.py
- DEP_Disable (requires NtSetInformationProcess; not sure if hooked already or not but I have not found a good exploit to test this with anyway): https://github.com/spender-sandbox/community-modified/blob/master/modules/signatures/dep_disable.py