Add in Locky ransomware signature
Converted cuckoo-modified signature: https://raw.githubusercontent.com/spender-sandbox/community-modified/master/modules/signatures/locky_apis.py. 

This has been tested and is working but like the pony signature that I submitted I had to remove this kind of stuff to replace with marking it as an IOC as no self.data now so unsure how to replicate " if tmp not in self.data". Still tested with recent locky samples and works fine.

                            for c2 in buf.group(1).split(","):
                                tmp = {"C2": c2}
                                if tmp not in self.data:
                                    self.data.append(tmp)