Fix up binary language sig
Made some changes as language codes do not appear in results anymore so the signature did not work. I have put some languages in but I have not managed to match up all the languages yet as while the list should be shorter now as we don't have to hit regional codes for an overall language some we are more interested in the sublangauge and not sure entirely how they are showing but I will fix them up. I have also tried to not "pick" on countries during this initial conversion of the signature but covered some of the main ones of more interest.

Also I have altered how the IOC is shown marking the entire PE resource so it can be seen from the signatures.