Create signature for network traffic for injection
This is a signature I am testing about system processes appearing in the process tree and generating network traffic which is possibly from code injection and seems to be the case in my testing. I also intend to do one with these appearing in the processes but not as a child process again indicative of injection without having to detect the injection itself.