Create signature for PDF static analysis
Added in signature to take advantage of peepdf although only JavaScript seems to currently seems there for features but will add more as things are added/I get detections in test although I did not have libemu on this system so I will sort that out. Is it the case that more peepdf stuff will be reported on (CVE, XFA extraction which is common in more recent exploits, suspicious features etc)?

Thanks.