Add in signature for crypto API use
This is more to begin highlighting how crypto APIs are being used by malware. Simply for now log it creating and exporting a key but may look as to how to highlight use in CnC & also file encryption. For now it is useful just to see if malware maybe using APIs versus stuff built into the malware code.