Update creates service sig
I noticed an FP issue when testing windows credential dumper where service name was not present (it was "None") so it said it wasn't started so I am moving to service handles instead which seem to be more reliable. Also I am markgin the starting of services as well so it becomes easier to see which services were and weren't started which is supicious as it can be used to delay malicious activity.