False Negative fix for Windows Utilities
I have found that using startswith results sometimes in a false negative if the command line starts with a path i.e C:\Windows\system32\netsh.exe advfirewall reset seen in cerber. I have put it back to utility in cmdline to fix this and to try and avoid false positives I have had to use syntax on some commands of "command " and "command.exe" where the change of the term appearing in an unrelated command is possible.