Add in signature for IRMA scan results
This is a basic sig I have written for IRMA scan results and probably can do with some more work but it does work it seems. I "trust" the scan results a lot more than virustotal in this case as people may have fewer scanners.

On another note if AV scanning done on every file regardless of OS shouldn't this and antivirus_virustotal.py be in cross instead of Windows folder?